+++ CHANGELOG.md
@@ -5,8 +5,12 @@ Tidarr notable changes.
 
 ## 📦 0.3.1
 ### 🖍 Changed
+* [Docker] ⚠️ **Now there is only one docker volume to mount for the music library destination**
+* [API] ⚠️ **Folder management and format is now handle in `tiddl.json` file.**
 * [API] Use tiddl `-V` parameters for video DL
 * [Front] Hide/Show video lists if tiddl `download_video` parameter is false/true.
+### 🚀 Added
+* [Front] Add an update warning modal
 ### 🐛 Fixed
 * [Front] Fix skeleton loaders